Move schema out of the validation registry
