+++ ludwig/schema/features/__init__.py
@@ -1,0 +1,13 @@
+import ludwig.schema.features.audio_feature
+import ludwig.schema.features.bag_feature
+import ludwig.schema.features.binary_feature
+import ludwig.schema.features.category_feature
+import ludwig.schema.features.date_feature
+import ludwig.schema.features.h3_feature
+import ludwig.schema.features.image_feature
+import ludwig.schema.features.number_feature
+import ludwig.schema.features.sequence_feature
+import ludwig.schema.features.set_feature
+import ludwig.schema.features.text_feature
+import ludwig.schema.features.timeseries_feature
+import ludwig.schema.features.vector_feature  # noqa
