+++ tests/ludwig/config_validation/test_validate_config_preprocessing.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ludwig.config_validation.validate_config import validate_config
+from ludwig.config_validation.validation import validate_config
 from ludwig.error import ConfigValidationError
 from tests.integration_tests.utils import binary_feature, category_feature
